gsheet-manipulation/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlibgspread
gspread-formatting
